internal: Improve config extending. (#274)
* Add extends extraction.

* Fix parsing.

* Move tests.

* More extends testing.

* Move configs.

* Move stuff around.

* Add more tests.

* Clean up fixtures.
