+++ crates/config/Cargo.toml
@@ -28,7 +28,9 @@ serde = { version = "1.0.140", features = ["derive"] }
 serde_json = { version = "1.0.82", features = ["preserve_order"] }
 serde_yaml = "0.9.4"
 strum = { version = "0.24.1", features = ["derive"] }
+thiserror = "1.0.31"
 validator = { version = "0.16.0", features = ["derive"] }
+yaml-rust = "0.4.5"
 
 [dev-dependencies]
 assert_fs = "1.0.7"
