--- crates/config/src/constants.rs
@@ -1,1 +1,0 @@
-pub const FLAG_PROJECTS_USING_GLOB: &str = "MOON_PROJECTS_USING_GLOBS";
