+++ crates/config/src/errors.rs
@@ -1,7 +1,37 @@
 use figment::{Error as FigmentError, Figment};
+use moon_error::MoonError;
 use serde_json::Value;
 use std::borrow::Cow;
+use std::path::PathBuf;
+use thiserror::Error;
 use validator::{ValidationError, ValidationErrors, ValidationErrorsKind};
+
+#[derive(Error, Debug)]
+pub enum ConfigError {
+    #[error("Failed validation.")]
+    FailedValidation(Vec<FigmentError>),
+
+    #[error("Invalid <id>extends</id> field, must be a string.")]
+    InvalidExtendsField,
+
+    #[error("Failed to parse YAML document <path>{0}</path>: {1}")]
+    InvalidYaml(PathBuf, String),
+
+    #[error("Cannot extend configuration file <file>{0}</file> as it does not exist.")]
+    MissingFile(String),
+
+    #[error("Unable to extend <file>{0}<file>, only YAML documents are supported.")]
+    UnsupportedExtendsDocument(String),
+
+    #[error("Cannot extend configuration file <file>{0}</file>, only HTTPS URLs are supported.")]
+    UnsupportedHttps(String),
+
+    #[error(transparent)]
+    Figment(#[from] FigmentError),
+
+    #[error(transparent)]
+    Moon(#[from] MoonError),
+}
 
 pub fn create_validation_error(code: &'static str, path: &str, message: String) -> ValidationError {
     let mut error = ValidationError::new(code);
@@ -10,12 +40,16 @@ pub fn create_validation_error(code: &'static str, path: &str, message: String) 
     error
 }
 
+pub fn format_error_line<T: AsRef<str>>(msg: T) -> String {
+    format!("  <accent>▪</accent> {}", msg.as_ref())
+}
+
 pub fn format_figment_errors(errors: Vec<FigmentError>) -> String {
     let mut list = vec![];
 
     for error in errors {
         for nested_error in error {
-            list.push(format!("  <accent>▪</accent> {}", nested_error));
+            list.push(format_error_line(nested_error.to_string()));
         }
     }
