+++ tests/fixtures/config-extends/project/global-0.yml
@@ -1,0 +1,5 @@
+tasks:
+  test:
+    command: 'jest'
+    inputs:
+      - 'tests'
