+++ tests/fixtures/config-extends/project/global-2.yml
@@ -1,0 +1,11 @@
+extends: './global-1.yml'
+
+fileGroups:
+  sources:
+    - 'sources/**/*'
+
+tasks:
+  format:
+    command: 'prettier'
+  test:
+    command: 'noop'
