+++ tests/fixtures/config-extends/workspace/base-0.yml
@@ -1,0 +1,6 @@
+actionRunner:
+  cacheLifetime: '3 hours'
+
+node:
+  version: '1.2.3'
+  addEnginesConstraint: true
