+++ tests/fixtures/config-extends/workspace/base-1.yml
@@ -1,0 +1,12 @@
+extends: './base-0.yml'
+
+actionRunner:
+  logRunningCommand: true
+
+node:
+  packageManager: 'yarn'
+  yarn:
+    version: '3.0.0'
+
+vcs:
+  manager: 'svn'
