+++ tests/fixtures/config-extends/workspace/base-2.yml
@@ -1,0 +1,8 @@
+extends: './base-1.yml'
+
+actionRunner:
+  logRunningCommand: false
+
+node:
+  version: '4.5.6'
+  dedupeOnLockfileChange: false
