+++ tests/fixtures/config-extends/workspace/typescript-0.yml
@@ -1,0 +1,2 @@
+# Not defined by default
+{}
