+++ tests/fixtures/config-extends/workspace/typescript-1.yml
@@ -1,0 +1,4 @@
+extends: './typescript-0.yml'
+
+typescript:
+  syncProjectReferences: false
