+++ tests/fixtures/config-extends/workspace/typescript-2.yml
@@ -1,0 +1,6 @@
+extends: './typescript-1.yml'
+
+typescript:
+  rootConfigFileName: 'tsconfig.root.json'
+  createMissingConfig: false
+  syncProjectReferences: true
