+++ tests/fixtures/config-invalid-global-project/.moon/workspace.yml
@@ -1,1 +1,3 @@
+extends: '../shared-workspace.yml'
+
 projects: {}
