+++ tests/fixtures/config-invalid-project/.moon/workspace.yml
@@ -1,2 +1,4 @@
+extends: '../shared-workspace.yml'
+
 projects:
   test: test
