+++ tests/fixtures/config-invalid-workspace/.moon/workspace.yml
@@ -1,5 +1,3 @@
-node:
-  # Use a unique version as to not collide with other tests
-  version: '16.1.0'
+extends: '../shared-workspace.yml'
 
 projects: true
