+++ tests/fixtures/migrate/.moon/workspace.yml
@@ -1,5 +1,4 @@
-node:
-  version: '16.0.0'
+extends: '../shared-workspace.yml'
 
 projects:
   - 'package-json/*'
