+++ tests/fixtures/project-graph/aliases/.moon/workspace.yml
@@ -1,6 +1,6 @@
+extends: '../shared-workspace.yml'
+
 node:
-  # Use a unique version as to not collide with other tests
-  version: '16.0.0'
   aliasPackageNames: 'name-and-scope'
 
 projects:
